# Improving the performance of the read_message by not splitting the da…